+++ changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Detection on `boxlang file` execution with invalid paths was not working properly.
+
 ## [1.0.0] - 2025-04-30
 
 [Unreleased]: https://github.com/ortus-boxlang/BoxLang/compare/v1.0.0...HEAD